Feedback

+++ src/fr/sciforma/apietnic/business/provider/ProjectFieldProvider.java
@@ -68,7 +68,7 @@ public class ProjectFieldProvider implements FieldProvider<Project> {
         fields.add(SciformaField.builder().name("Finish Constraint").type(FieldType.DATE).build());
         fields.add(SciformaField.builder().name("ID").type(FieldType.STRING).build());
         fields.add(SciformaField.builder().name("Initiator").type(FieldType.USER).build());
-        fields.add(SciformaField.builder().name("Internal ID").type(FieldType.DECIMAL).build());
+        fields.add(SciformaField.builder().name("Internal ID").type(FieldType.INTEGER).build());
         fields.add(SciformaField.builder().name("Manager 1").type(FieldType.USER).build());
         fields.add(SciformaField.builder().name("Manager 2").type(FieldType.USER).build());
         fields.add(SciformaField.builder().name("Manager 3").type(FieldType.USER).build());
