+++ src/fr/sciforma/apietnic/business/provider/UserFieldProvider.java
@@ -22,7 +22,7 @@ public class UserFieldProvider implements FieldProvider<User> {
         fields.add(SciformaField.builder().name("Email Address 1").type(FieldType.STRING).build());
         fields.add(SciformaField.builder().name("First Name").type(FieldType.STRING).build());
         fields.add(SciformaField.builder().name("ID").type(FieldType.STRING).build());
-        fields.add(SciformaField.builder().name("Internal ID").type(FieldType.DECIMAL).build());
+        fields.add(SciformaField.builder().name("Internal ID").type(FieldType.INTEGER).build());
         fields.add(SciformaField.builder().name("Last Name").type(FieldType.STRING).build());
         fields.add(SciformaField.builder().name("Login ID").type(FieldType.STRING).build());
         fields.add(SciformaField.builder().name("Middle Name").type(FieldType.STRING).build());
